refactor: Moved format util file
