+++ @navikt/core/tokens/darkside/format.ts
